+++ example/jenkins-cli/src/Config.hs
@@ -11,12 +11,12 @@ import           Control.Lens
 import           Control.Applicative
 import           Data.Aeson
 import qualified Data.ByteString.Lazy as ByteString
-import           Jenkins.Rest
+import qualified Jenkins.Rest as Jenkins
 import           System.Directory (getAppUserDataDirectory)
 import           System.FilePath ((</>))
 
 
-newtype Config = Config { _unConfig :: Master }
+newtype Config = Config { _unConfig :: Jenkins.Master }
   deriving (Show, Eq)
 
 instance FromJSON Config where
@@ -24,13 +24,13 @@ instance FromJSON Config where
     url   <- o .: "url"
     user  <- o .: "user"
     token <- o .: "api-token"
-    return (Config (defaultMaster
-      & jenkinsUrl .~ url
-      & jenkinsUser .~ user
-      & jenkinsApiToken .~ token))
+    return (Config (Jenkins.defaultMaster
+      & set Jenkins.url url
+      & set Jenkins.user user
+      & set Jenkins.apiToken token))
   parseJSON _ = empty
 
-readConfig :: IO Master
+readConfig :: IO Jenkins.Master
 readConfig = do
   appData <- getAppUserDataDirectory "jenkins-cli"
   either error _unConfig . eitherDecode <$> ByteString.readFile (appData </> "conf.json")