+++ test/Jenkins/RestSpec.hs
@@ -12,7 +12,7 @@ import qualified Data.Conduit as C
 import           Data.Functor.Identity (Identity)
 import           Data.Monoid (mempty)
 import           Test.Hspec
-import qualified Jenkins.Rest as Rest
+import qualified Jenkins.Rest as Jenkins
 import           Jenkins.Rest.Internal
 import qualified Jenkins.Rest.Method.Internal as Method
 
@@ -22,25 +22,25 @@ spec = do
   context "POST requests" $ do
     it "post_ sends POST request with empty body" $ do
       interpret $ do
-        Rest.post_ "foo"
-        Rest.post_ "bar"
-        Rest.post_ "baz"
+        Jenkins.post_ "foo"
+        Jenkins.post_ "bar"
+        Jenkins.post_ "baz"
      `shouldBe`
       [QPost 0 "" "foo", QPost 1 "" "bar", QPost 2 "" "baz"]
 
     it "post sends POST request with non-empty body" $ do
       interpret $ do
-        Rest.post "foo" "qux"
-        Rest.post "bar" "quux"
-        Rest.post "baz" "xyzzy"
+        Jenkins.post "foo" "qux"
+        Jenkins.post "bar" "quux"
+        Jenkins.post "baz" "xyzzy"
      `shouldBe`
       [QPost 0 "qux" "foo", QPost 1 "quux" "bar", QPost 2 "xyzzy" "baz"]
 
   context "GET requests" $
     it "get sends GET requests" $ do
       interpret $ do
-        let getS :: (forall f. Rest.Method Rest.Complete f) -> JenkinsT Identity (ResourceT IO (C.ResumableSource (ResourceT IO) Strict.ByteString))
-            getS = Rest.getS Rest.plain
+        let getS :: (forall f. Jenkins.Method Jenkins.Complete f) -> JenkinsT Identity (ResourceT IO (C.ResumableSource (ResourceT IO) Strict.ByteString))
+            getS = Jenkins.getS Jenkins.plain
         getS "foo"
         getS "bar"
         getS "baz"
@@ -51,24 +51,24 @@ spec = do
   describe "reload" $
     it "calls $jenkins_url/reload with POST query and then disconnects" $ do
       interpret $ do
-        Rest.reload
-        Rest.post_ "foo"
+        Jenkins.reload
+        Jenkins.post_ "foo"
      `shouldBe`
       [QPost 0 "" "reload", QDisconnect]
 
   describe "restart" $
     it "calls $jenkins_url/safeRestart with POST query and then disconnects" $ do
       interpret $ do
-        Rest.restart
-        Rest.post_ "bar"
+        Jenkins.restart
+        Jenkins.post_ "bar"
      `shouldBe`
       [QPost 0 "" "safeRestart", QDisconnect]
 
   describe "forceRestart" $
     it "calls $jenkins_url/restart with POST query and then disconnects" $ do
       interpret $ do
-        Rest.forceRestart
-        Rest.post_ "baz"
+        Jenkins.forceRestart
+        Jenkins.post_ "baz"
      `shouldBe`
       [QPost 0 "" "restart", QDisconnect]
 
@@ -82,9 +82,9 @@ data Query =
 newtype Requests a = Requests [a]
   deriving (Show, Eq)
 
-interpret :: Rest.JenkinsT Identity a -> [Query]
-interpret adt = evalState (iterJenkins go ([] <$ adt)) (Requests [0..]) where
-  go :: JenkinsF Identity (State (Requests Int) [Query]) -> State (Requests Int) [Query]
+interpret :: JenkinsT Identity a -> [Query]
+interpret adt = evalState (iter go ([] <$ adt)) (Requests [0..]) where
+  go :: JF Identity (State (Requests Int) [Query]) -> State (Requests Int) [Query]
   go (Get m n) = do
     r <- render QGet m
     fmap (r :) (n (return (C.newResumableSource (C.yield mempty))))
@@ -94,7 +94,7 @@ interpret adt = evalState (iterJenkins go ([] <$ adt)) (Requests [0..]) where
   go Dcon =
     return [QDisconnect]
 
-render :: (a -> Strict.ByteString -> Query) -> Rest.Method Method.Complete f -> State (Requests a) Query
+render :: (a -> Strict.ByteString -> Query) -> Jenkins.Method Method.Complete f -> State (Requests a) Query
 render f m = do
   n <- next
   return $ f n (Method.render m)